requirements.txt
apify-client>=1.6.0
pandas>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
pyyaml>=6.0

